Refatornado o LESS do panels/_news.less
