+++ web/src/js/profile.js
@@ -1,6 +1,6 @@
 import HTMLLoaderMenu from './components/html-loader-menu.js';
 import Switch from './components/switch-input.js';
-import newsAction from './view/profile-news.js';
+import newsAction from './view/news.js';
 import validateSession from './helpers/session-validator.js';
 import GoogleLogin from './helpers/google-login.js';
