+++ web/src/less/profile.less
@@ -1,9 +1,10 @@
+@import "vars.less";
+@import "mixins/all.less";
+@import "components/all.less";
+@import "resets.less";
+@import "panels/all.less";
 @import "aside.less";
 @import "content.less";
 @import "header.less";
 @import "medias.less";
-@import "resets.less";
 @import "toast.less";
-@import "panels/all.less";
-@import "vars.less";
-@import "mixins/all.less";
